atomic_move - fix preserving extended acls (#82818)
Based on 2bb09bfd12e42e7be6f39ab0e45a992512c240f9, but compatible with
Python 2.7 and the 2.16 CI matrix.

Fix preserving extended attributes for Python 3 by using copystat

For Python 2, parse getfacl ouptut for POSIX ACLs and copy extended
attributes using setfacl, ignoring errors

update unit test mocks for added method of attribute preservation

remove erroneous `- meta: end_play` from lineinfile test suite

add integration test for lineinfile case

modify integration test - openSUSE needs the user to exist

fix nested template in copy test assertion

ci_complete

+++ lib/ansible/module_utils/basic.py
@@ -246,10 +246,12 @@ MODE_OPERATOR_RE = re.compile(r'[+=-]')
 USERS_RE = re.compile(r'^[ugo]+$')
 PERMS_RE = re.compile(r'^[rwxXstugo]*$')
 
+POSIX_ACL_RE = re.compile(r'^((?:user|group|other):[\S]*:[r-][w-][x-])[^\n\r]*$', flags=re.M)
 
 #
 # Deprecated functions
 #
+
 
 def get_platform():
     '''
@@ -1648,7 +1650,21 @@ class AnsibleModule(object):
         current_attribs = current_attribs.get('attr_flags', '')
         self.set_attributes_if_different(dest, current_attribs, True)
 
-    def atomic_move(self, src, dest, unsafe_writes=False):
+    def _copystat_ea_best_effort(self, acl_origin, acl_dest):
+        shutil.copystat(acl_origin, acl_dest)
+        if PY3 or platform.system() != 'Linux':
+            return
+        getfacl = self.get_bin_path('getfacl')
+        setfacl = self.get_bin_path('setfacl')
+        if getfacl is None or setfacl is None:
+            return
+        rc, out, err = self.run_command("{getfacl} {path}".format(getfacl=getfacl, path=acl_origin))
+        if rc != 0:
+            return
+        for acl in POSIX_ACL_RE.findall(out):
+            self.run_command("{setfacl} --modify={acl} {path}".format(setfacl=setfacl, acl=acl, path=acl_dest))
+
+    def atomic_move(self, src, dest, unsafe_writes=False, keep_dest_attrs=True):
         '''atomically move src to dest, copying attributes from dest, returns true on success
         it uses os.rename to ensure this as it is an atomic operation, rest of the function is
         to work around limitations, corner cases and ensure selinux context is saved if possible'''
@@ -1656,24 +1672,16 @@ class AnsibleModule(object):
         dest_stat = None
         b_src = to_bytes(src, errors='surrogate_or_strict')
         b_dest = to_bytes(dest, errors='surrogate_or_strict')
-        if os.path.exists(b_dest):
+        if os.path.exists(b_dest) and keep_dest_attrs:
             try:
                 dest_stat = os.stat(b_dest)
 
-                # copy mode and ownership
-                os.chmod(b_src, dest_stat.st_mode & PERM_BITS)
+                # copy ownership
                 os.chown(b_src, dest_stat.st_uid, dest_stat.st_gid)
 
-                # try to copy flags if possible
-                if hasattr(os, 'chflags') and hasattr(dest_stat, 'st_flags'):
-                    try:
-                        os.chflags(b_src, dest_stat.st_flags)
-                    except OSError as e:
-                        for err in 'EOPNOTSUPP', 'ENOTSUP':
-                            if hasattr(errno, err) and e.errno == getattr(errno, err):
-                                break
-                        else:
-                            raise
+                # copy any attributes possible
+                self._copystat_ea_best_effort(b_dest, b_src)
+
             except OSError as e:
                 if e.errno != errno.EPERM:
                     raise
@@ -1721,18 +1729,29 @@ class AnsibleModule(object):
                             os.close(tmp_dest_fd)
                             # leaves tmp file behind when sudo and not root
                             try:
-                                shutil.move(b_src, b_tmp_dest_name)
+                                if PY3:
+                                    shutil.move(b_src, b_tmp_dest_name, copy_function=shutil.copy if keep_dest_attrs else shutil.copy2)
+                                elif not keep_dest_attrs:
+                                    shutil.move(b_src, b_tmp_dest_name)
+                                else:
+                                    shutil.copy(b_src, b_tmp_dest_name)
                             except OSError:
                                 # cleanup will happen by 'rm' of tmpdir
                                 # copy2 will preserve some metadata
-                                shutil.copy2(b_src, b_tmp_dest_name)
+                                if not keep_dest_attrs:
+                                    shutil.copy2(b_src, b_tmp_dest_name)
+                                elif PY3:
+                                    shutil.copy(b_src, b_tmp_dest_name)
+
+                            if not PY3 and not keep_dest_attrs:
+                                self._copystat_ea_best_effort(b_src, b_tmp_dest_name)
 
                             if self.selinux_enabled():
                                 self.set_context_if_different(
                                     b_tmp_dest_name, context, False)
                             try:
                                 tmp_stat = os.stat(b_tmp_dest_name)
-                                if dest_stat and (tmp_stat.st_uid != dest_stat.st_uid or tmp_stat.st_gid != dest_stat.st_gid):
+                                if keep_dest_attrs and dest_stat and (tmp_stat.st_uid != dest_stat.st_uid or tmp_stat.st_gid != dest_stat.st_gid):
                                     os.chown(b_tmp_dest_name, dest_stat.st_uid, dest_stat.st_gid)
                             except OSError as e:
                                 if e.errno != errno.EPERM:
